requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Django==3.2
djangorestframework==3.13.1
idna==3.3
joblib==1.1.0
nltk==3.6.5
numpy==1.21.4
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
regex==2021.11.10
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
urllib3==1.26.7
